requirements.txt
boto3==1.4.7
botocore==1.7.47
cffi==1.11.2
Django==1.11.7
docutils==0.14
file==0.3.0
futures==3.1.1
isort==4.2.15
jmespath==0.9.3
pycparser==2.18
python-dateutil==2.6.1
python-magic==0.4.13
pytz==2017.3
s3transfer==0.1.11
six==1.11.0
